refactor(orchestra): extract auto-scene recovery into dedicated service
Split 180+ lines of damaged-worktree recovery logic out of coordinator.py
into src/vibe3/execution/auto_scene_recovery.py to bring coordinator back
under the 400-line limit (590 -> 417). Moved corresponding tests to
test_auto_scene_recovery.py.
