+++ tests/vibe3/execution/test_coordinator.py
@@ -9,7 +9,6 @@ import pytest
 from vibe3.execution.contracts import ExecutionRequest
 from vibe3.execution.coordinator import ExecutionCoordinator
 from vibe3.execution.role_contracts import WorktreeRequirement
-from vibe3.models.orchestration import IssueState
 
 
 @pytest.fixture
@@ -248,144 +247,6 @@ def test_sync_child_clears_async_marker_before_entering_sync_shell(
     assert result.launched is True
     assert seen_async_marker == [None]
     assert os.environ.get("VIBE3_ASYNC_CHILD") == "1"
-
-
-def test_auto_scene_reset_recovers_damaged_auto_worktree(
-    mock_dependencies,
-    tmp_path: Path,
-):
-    """Detached registered auto scenes should be reset and re-queued."""
-    config, store, backend, capacity = mock_dependencies
-    capacity.can_dispatch.return_value = True
-
-    target_path = tmp_path / ".worktrees" / "task/issue-42"
-    target_path.mkdir(parents=True)
-
-    coordinator = ExecutionCoordinator(
-        config=config,
-        store=store,
-        backend=backend,
-        capacity=capacity,
-    )
-    coordinator.registry.get_truly_live_sessions_for_branch = MagicMock(return_value=[])
-
-    request = ExecutionRequest(
-        role="manager",
-        target_branch="task/issue-42",
-        target_id=42,
-        execution_name="vibe3-manager-issue-42",
-        cmd=["echo", "hello"],
-        repo_path=str(tmp_path),
-        worktree_requirement=WorktreeRequirement.PERMANENT,
-        tick_id=7,
-    )
-
-    with (
-        patch.object(
-            coordinator,
-            "_resolve_cwd",
-            side_effect=ValueError(
-                "Failed to resolve permanent worktree for manager:42"
-            ),
-        ),
-        patch(
-            "vibe3.environment.worktree_support.find_worktree_by_path",
-            return_value=True,
-        ),
-        patch(
-            "vibe3.environment.worktree_support.find_worktree_for_branch",
-            return_value=None,
-        ),
-        patch.object(coordinator, "_read_worktree_head", return_value="HEAD"),
-        patch("vibe3.exceptions.error_tracking.ErrorTrackingService") as mock_tracking,
-        patch(
-            "vibe3.services.flow_cleanup_service.FlowCleanupService"
-        ) as mock_cleanup_cls,
-        patch("vibe3.services.label_service.LabelService") as mock_label_cls,
-    ):
-        mock_cleanup = MagicMock()
-        mock_cleanup.cleanup_flow_scene.return_value = {
-            "worktree": True,
-            "local_branch": True,
-            "remote_branch": True,
-            "handoff": True,
-            "flow_record": True,
-        }
-        mock_cleanup_cls.return_value = mock_cleanup
-        mock_label = MagicMock()
-        mock_label_cls.return_value = mock_label
-
-        result = coordinator.dispatch_execution(request)
-
-    assert result.launched is False
-    assert result.skipped is True
-    assert result.reason_code == "auto_scene_reset"
-    mock_tracking.get_instance.return_value.record_error.assert_called_once()
-    mock_cleanup.cleanup_flow_scene.assert_called_once_with(
-        "task/issue-42",
-        include_remote=True,
-        terminate_sessions=True,
-        keep_flow_record=False,
-    )
-    mock_label.confirm_issue_state.assert_called_once_with(
-        42,
-        IssueState.READY,
-        actor="orchestra:auto-recover",
-        force=True,
-    )
-    assert store.add_event.call_count == 2
-
-
-def test_auto_scene_reset_skips_when_live_session_exists(
-    mock_dependencies,
-    tmp_path: Path,
-):
-    """Live auto scenes must not be force-reset by the recovery path."""
-    config, store, backend, capacity = mock_dependencies
-    capacity.can_dispatch.return_value = True
-
-    target_path = tmp_path / ".worktrees" / "task/issue-42"
-    target_path.mkdir(parents=True)
-
-    coordinator = ExecutionCoordinator(
-        config=config,
-        store=store,
-        backend=backend,
-        capacity=capacity,
-    )
-    coordinator.registry.get_truly_live_sessions_for_branch = MagicMock(
-        return_value=[{"id": 1, "branch": "task/issue-42", "tmux_session": "live"}]
-    )
-
-    request = ExecutionRequest(
-        role="manager",
-        target_branch="task/issue-42",
-        target_id=42,
-        execution_name="vibe3-manager-issue-42",
-        cmd=["echo", "hello"],
-        repo_path=str(tmp_path),
-        worktree_requirement=WorktreeRequirement.PERMANENT,
-    )
-
-    with (
-        patch.object(
-            coordinator,
-            "_resolve_cwd",
-            side_effect=ValueError(
-                "Failed to resolve permanent worktree for manager:42"
-            ),
-        ),
-        patch("vibe3.exceptions.error_tracking.ErrorTrackingService") as mock_tracking,
-        patch(
-            "vibe3.services.flow_cleanup_service.FlowCleanupService"
-        ) as mock_cleanup_cls,
-    ):
-        result = coordinator.dispatch_execution(request)
-
-    assert result.launched is False
-    assert result.reason_code == "worktree_unavailable"
-    mock_tracking.get_instance.return_value.record_error.assert_not_called()
-    mock_cleanup_cls.assert_not_called()
 
 
 def test_sync_skips_when_live_session_exists_for_target(
